Add deleting categories and accounts and getting list of them

+++ main.py
@@ -17,40 +17,14 @@ import messages
 nest_asyncio.apply()
 
 class UsersData:
-    """Uses async SQLite to storage and manipulate data of users. (Singleton)
-
-    Methods
-
-    init():
-        initializes (connects to) a database
-    create():
-        creates a database if it doesn't exist
-    delete():
-        deletes a database if it exists
-    add_user(user_id):
-        adds a user to database if it was not added before
-    set_value(user_id, name, new_value):
-        changes value of field 'name' of user 'user_id' to 'new_value'
-    add_value(user_id, name, delta):
-        adds 'delta' to value 'name' of user 'user_id'
-    get_value(name, value, result):
-        returns value of 'result' of user with 'name' = 'value'
-    exists(name, value):
-        returns user_id of user with 'name' = 'value', 0 otherwise
-    """
-
     _cur = None
 
     async def init(self, connection):
-        """Initializes and creates database (or connects if it already exists)"""
-
         connection.row_factory = sq.Row
         self._cur = await connection.cursor()
         await self.create()
 
     async def create(self):
-        """Creates new database if it doesn't exists"""
-
         await self._cur.execute("""CREATE TABLE IF NOT EXISTS users (
             user_id INTEGER PRIMARY KEY
             )""")
@@ -75,98 +49,10 @@ class UsersData:
             )""")
 
     async def delete(self):
-        """Deletes database if it exists"""
-
         await self._cur.execute("""DROP TABLE IF EXISTS users""")
         await self._cur.execute("""DROP TABLE IF EXISTS accounts""")
         await self._cur.execute("""DROP TABLE IF EXISTS categories""")
         await self._cur.execute("""DROP TABLE IF EXISTS transactions""")
-
-    async def add_user(self, user_id):
-        """Adds user to database
-        Raises DoubleInitError if user already exists
-        """
-
-        try:
-            await self._cur.execute("""INSERT INTO users (user_id) VALUES (?)""", (user_id,))
-            return True
-        except Exception:
-            return False
-
-    async def set_value(self, user_id, name, new_value):
-        """Changes value of 'name' of user 'user_id' to 'new_value'
-        Creates new user if it is not added to database
-        """
-
-        try:
-            await self._cur.execute("""INSERT INTO users (user_id, ?) 
-                VALUES (?, ?)""",
-                (name, user_id, new_value))
-        except Exception:
-            await self._cur.execute(f"""UPDATE users SET {name} = {new_value} 
-                WHERE user_id = {user_id}""")
-
-    async def add_value(self, user_id, name, delta):
-        """Adds 'delta' to value 'name' of user 'user_id'
-        Creates new user if it is not added to database
-        """
-
-        old_value = None
-        try: 
-            old_value = await self.get_value("user_id", user_id, name)
-        except KeyError:
-            old_value = 0 
-
-        await self.set_value(user_id, name, old_value + delta)
-
-    async def get_value(self, name, value, result):
-        """Returns value of 'result' of user with 'name' = 'value'
-        Raises KeyError if not found
-        """
-
-        await self._cur.execute(f"""SELECT {result} FROM users 
-            WHERE {name} = {value}""")
-
-        result = await self._cur.fetchone()
-        if result is None:
-            raise KeyError
-        return result[0]
-
-    async def exists_user(self, user_id):
-        await self._cur.execute("""SELECT * FROM users WHERE user_id = ?""", (user_id,))
-
-        result = await self._cur.fetchone()
-        if result is None:
-            return False
-        return True
-
-    async def get_category_id(self, user_id, category_name):
-        await self._cur.execute("""SELECT category_id FROM categories
-                WHERE (owner_id, name) = (?, ?)""",
-                (user_id, category_name))
-
-        result = await self._cur.fetchone()
-        if result is None:
-            return -1
-        return result[0]
-
-    async def get_account_id(self, user_id, account_name):
-        await self._cur.execute("""SELECT account_id FROM accounts
-                WHERE (owner_id, name) = (?, ?)""",
-                (user_id, account_name))
-
-        result = await self._cur.fetchone()
-        if result is None:
-            return -1
-        return result[0]
-
-    async def exists_category(self, user_id, category_name):
-        result = await self.get_category_id(user_id, category_name)
-        return result != -1
-
-    async def exists_account(self, user_id, account_name):
-        result = await self.get_account_id(user_id, account_name)
-        return result != -1
 
     async def get_balance(self, user_id, account_name=""):
         if account_name:
@@ -183,31 +69,96 @@ class UsersData:
         result = await self._cur.fetchone()
         return result[1] if result[0] else 0
 
-    async def add_category(self, user_id, name):
-        exists = await self.exists_category(user_id, name)
-        if exists:
-            return False
+    async def get_category_id(self, user_id, category_name):
+        await self._cur.execute("""SELECT category_id FROM categories
+                WHERE (owner_id, name) = (?, ?)""",
+                (user_id, category_name))
 
+        result = await self._cur.fetchone()
+        return -1 if result is None else result[0]
+
+    async def get_account_id(self, user_id, account_name):
+        await self._cur.execute("""SELECT account_id FROM accounts
+                WHERE (owner_id, name) = (?, ?)""",
+                (user_id, account_name))
+
+        result = await self._cur.fetchone()
+        return -1 if result is None else result[0]
+
+    async def exists_user(self, user_id):
+        await self._cur.execute("""SELECT * FROM users WHERE user_id = ?""", (user_id,))
+
+        result = await self._cur.fetchone()
+        return result is not None
+
+    async def exists_category(self, user_id, category_name):
+        result = await self.get_category_id(user_id, category_name)
+        return result != -1
+
+    async def exists_account(self, user_id, account_name):
+        result = await self.get_account_id(user_id, account_name)
+        return result != -1
+
+    async def add_user(self, user_id):
+        await self._cur.execute("""INSERT INTO users (user_id) VALUES (?)""", (user_id,))
+
+    async def add_category(self, user_id, name):
         await self._cur.execute("""INSERT INTO categories (owner_id, name) 
                 VALUES (?, ?)""",
                 (user_id, name))
-        return True
 
     async def add_account(self, user_id, name):
-        exists = await self.exists_account(user_id, name)
-        if exists:
-            return False
-
         await self._cur.execute("""INSERT INTO accounts (owner_id, name) 
                 VALUES (?, ?)""",
                 (user_id, name))
-        return True
 
     async def add_transaction(self, user_id, amount, category_id, account_id):
         await self._cur.execute("""INSERT INTO transactions 
                 (category_id, account_id, amount)
                 VALUES (?, ?, ?)""",
                 (category_id, account_id, amount))
+
+    async def delete_category(self, user_id, category_id):
+        await self._cur.execute("""DELETE FROM transactions
+            WHERE category_id = ?""",
+            (category_id,))
+
+        await self._cur.execute("""DELETE FROM categories 
+                WHERE category_id = ?""",
+                (category_id,))
+
+    async def delete_account(self, user_id, account_id):
+        await self._cur.execute("""DELETE FROM transactions
+            WHERE account_id = ?""",
+            (account_id,))
+
+        await self._cur.execute("""DELETE FROM accounts 
+                WHERE account_id = ?""",
+                (account_id,))
+
+    async def get_categories(self, user_id):
+        await self._cur.execute("""SELECT categories.name FROM categories
+                WHERE owner_id = ?""",
+                (user_id,))
+
+        categories = []
+        result = await self._cur.fetchall()
+        for row in result:
+            categories.append(row[0])
+        return categories
+
+    async def get_accounts(self, user_id):
+        await self._cur.execute("""SELECT accounts.name FROM accounts
+                WHERE owner_id = ?""",
+                (user_id,))
+
+        accounts = []
+        result = await self._cur.fetchall()
+        for row in result:
+            accounts.append(row[0])
+        return accounts
+
+
 
 
 
@@ -231,53 +182,29 @@ async def main():
 
             @dp.message_handler(commands=["start"])
             async def start_message(message: types.Message):
-                await message.reply(messages.START)
-
-            @dp.message_handler(commands=["get_id"])
-            async def get_id_message(message: types.Message):
                 user_id = message.from_user.id
+                result = await data.exists_user(user_id)
 
-                await message.reply(messages.GET_ID.format(user_id=user_id), 
-                        parse_mode=types.ParseMode.HTML)
+                if not result:
+                    result = await data.add_user(user_id)
 
-
-            @dp.message_handler(commands=["register"])
-            async def registration_message(message: types.Message):
-                user_id = message.from_user.id
-
-                result = await data.add_user(user_id)
-
-                if result:
-                    tg.create_task(bot.send_message(message.from_user.id,
-                            messages.SUCCESSFUL_REGISTRATION))
-                else:
-                    tg.create_task(bot.send_message(message.from_user.id,
-                            messages.DOUBLE_REGISTRATION))
-
+                tg.create_task(message.reply(messages.START))
                 tg.create_task(con.commit())
 
-            @dp.message_handler(commands=["recreate"])
+            @dp.message_handler(commands=["recreate", "кускуфеу"])
             async def recreate_message(message: types.Message):
-                await data.delete()
-                await data.create()
+                if message.from_user.id == config.ADMIN_ID:
+                    await data.delete()
+                    await data.create()
 
-                tg.create_task(bot.send_message(message.from_user.id, messages.RECREATE_DONE,
-                        reply_markup=ReplyKeyboardRemove()))
-                tg.create_task(con.commit())
+                    tg.create_task(bot.send_message(message.from_user.id, messages.RECREATE_DONE,
+                            reply_markup=ReplyKeyboardRemove()))
+                    tg.create_task(con.commit())
 
 
-            @dp.message_handler(commands=["help"])
+            @dp.message_handler(commands=["help", "рудз"])
             async def help_message(message: types.Message):
-                kb = [
-                    [KeyboardButton(text="/get_id")],
-                    [KeyboardButton(text="/recreate")],
-                    [KeyboardButton(text="/register")],
-                    [KeyboardButton(text="баланс")]
-                ]
-                keyboard = ReplyKeyboardMarkup(keyboard=kb)
-
-                await message.reply(messages.HELP, reply_markup=keyboard,
-                        parse_mode=types.ParseMode.HTML)
+                await message.reply(messages.HELP, parse_mode=types.ParseMode.HTML)
 
             @dp.message_handler()
             async def message_reply(message: types.Message):
@@ -291,25 +218,55 @@ async def main():
 
                 match tokens:
                     case "добавить"|"доб", "категорию"|"кат", name:
-                        result = await data.add_category(user_id, name)
-                        if not result:
+                        exists = await data.exists_category(user_id, name)
+
+                        if exists:
                             tg.create_task(message.reply(
                                     messages.DOUBLE_CATEGORY_ADD.format(name=name),
                                     parse_mode=types.ParseMode.HTML))
                         else:
+                            await data.add_category(user_id, name)
                             tg.create_task(message.reply(
                                     messages.SUCCESSFUL_CATEGORY_ADD.format(name=name),
                                     parse_mode=types.ParseMode.HTML))
 
                     case "добавить"|"доб", "счет"|"счёт", name:
-                        result = await data.add_account(user_id, name)
-                        if not result:
+                        exists = await data.exists_account(user_id, name)
+
+                        if exists:
                             tg.create_task(message.reply(
                                     messages.DOUBLE_ACCOUNT_ADD.format(name=name),
                                     parse_mode=types.ParseMode.HTML))
                         else:
+                            result = await data.add_account(user_id, name)
                             tg.create_task(message.reply(
                                     messages.SUCCESSFUL_ACCOUNT_ADD.format(name=name),
+                                    parse_mode=types.ParseMode.HTML))
+
+                    case "удалить"|"уд", "категорию"|"кат", name:
+                        category_id = await data.get_category_id(user_id, name)
+
+                        if category_id == -1:
+                            tg.create_task(message.reply(
+                                    messages.CATEGORY_NOT_EXIST.format(name=name),
+                                    parse_mode=types.ParseMode.HTML))
+                        else:
+                            await data.delete_category(user_id, category_id)
+                            tg.create_task(message.reply(
+                                    messages.SUCCESSFUL_CATEGORY_DELETE.format(name=name),
+                                    parse_mode=types.ParseMode.HTML))
+
+                    case "удалить"|"уд", "счет"|"счёт", name:
+                        account_id = await data.get_account_id(user_id, name)
+
+                        if account_id == -1:
+                            tg.create_task(message.reply(
+                                    messages.ACCOUNT_NOT_EXIST.format(name=name),
+                                    parse_mode=types.ParseMode.HTML))
+                        else:
+                            await data.delete_account(user_id, account_id)
+                            tg.create_task(message.reply(
+                                    messages.SUCCESSFUL_ACCOUNT_DELETE.format(name=name),
                                     parse_mode=types.ParseMode.HTML))
 
                     case amount, category, account if isfloat(amount.replace(',', '.')):
@@ -356,6 +313,30 @@ async def main():
                         tg.create_task(message.reply(
                                 messages.BALANCE.format(balance=round(balance, 2)),
                                 parse_mode=types.ParseMode.HTML))
+
+                    case "категории",:
+                        categories = await data.get_categories(user_id)
+
+                        if not categories:
+                            tg.create_task(message.reply(messages.NO_CATEGORIES,
+                                    parse_mode=types.ParseMode.HTML))
+                        else:
+                            tg.create_task(message.reply(
+                                    messages.CATEGORIES.format(
+                                        categories=messages.CATEGORIES_SEP.join(categories)),
+                                    parse_mode=types.ParseMode.HTML))
+
+                    case "счета",:
+                        accounts = await data.get_accounts(user_id)
+
+                        if not accounts:
+                            tg.create_task(message.reply(messages.NO_ACCOUNTS,
+                                    parse_mode=types.ParseMode.HTML))
+                        else:
+                            tg.create_task(message.reply(
+                                    messages.ACCOUNTS.format(
+                                        accounts=messages.ACCOUNTS_SEP.join(accounts)),
+                                    parse_mode=types.ParseMode.HTML))
 
                     case _:
                         tg.create_task(message.reply(messages.UNKNOWN_COMMAND))
